Move ready_services to tower-ready-cache
