+++ tower-balance/Cargo.toml
@@ -24,7 +24,6 @@ publish = false
 
 [dependencies]
 futures = "0.1.26"
-indexmap = "1.0.2"
 log = "0.4.1"
 rand = "0.6.5"
 tokio-sync = "0.1.3"
@@ -32,6 +31,7 @@ tokio-timer = "0.2.4"
 tower-discover = "0.1.0"
 tower-layer = "0.1.0"
 tower-load = { version = "0.1.0", path = "../tower-load" }
+tower-ready-cache = { version = "0.1.0", path = "../tower-ready-cache" }
 tower-service = "0.2.0"
 tower-util = "0.1.0"
