+++ tower-ready-cache/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "tower-ready-cache"
+# When releasing to crates.io:
+# - Remove path dependencies
+# - Update html_root_url.
+# - Update doc url
+#   - Cargo.toml
+#   - README.md
+# - Update CHANGELOG.md.
+# - Create "v0.1.x" git tag.
+version = "0.1.0"
+authors = ["Tower Maintainers <team@tower-rs.com>"]
+license = "MIT"
+readme = "README.md"
+repository = "https://github.com/tower-rs/tower"
+homepage = "https://github.com/tower-rs/tower"
+documentation = "https://docs.rs/tower-ready-cache/0.1.0"
+description = """
+Caches a set of services
+"""
+categories = ["asynchronous", "network-programming"]
+edition = "2018"
+publish = false
+
+[dependencies]
+futures = "0.1.26"
+indexmap = "1.0.2"
+log = "0.4.1"
+tokio-sync = "0.1.3"
+tower-service = "0.2.0"
+tower-util = "0.1.0"
